pythonV/requirements.txt
asgiref==3.11.1
Django==4.2.11
psycopg2-binary==2.9.11
sqlparse==0.5.5
tzdata==2025.3
